+++ src/aprl/agents/monte_carlo.py
@@ -6,83 +6,7 @@ from multiprocessing import Pipe, Process
 import gym
 from stable_baselines.common.vec_env import CloudpickleWrapper
 
-from aprl.common.mujoco import MujocoState
-from aprl.envs.multi_agent import MultiWrapper
-
-
-class ResettableEnv(gym.Env):
-    """A Gym environment that can be reset to an arbitrary state."""
-    @abstractmethod
-    def get_state(self):
-        """Returns a serialized representation of the current state."""
-        pass
-
-    @abstractmethod
-    def set_state(self, x):
-        """Restores the environment to a previously saved state.
-        :param x: return value of a previous call to get_state()."""
-        pass
-
-
-class OldMujocoResettableWrapper(ResettableEnv, MultiWrapper):
-    """Converts a MujocoEnv into a ResettableEnv.
-
-    Note all MuJoCo environments are resettable."""
-    def __init__(self, env):
-        """Wraps a MujocoEnv, adding get_state and set_state methods.
-        :param env: a MujocoEnv. NOTE: it must not be wrapped in a TimeLimit."""
-        if hasattr(env, '_max_episode_steps'):
-            raise TypeError('Environment must not have a time limit '
-                            '(try passing in env.unwrapped instead).')
-        gym.Wrapper.__init__(self, env)
-        self.sim = env.unwrapped.env_scene
-
-    def get_full_state(self):
-        state_dict = {}
-        for k, v in type(self.sim.data._wrapped.contents).__dict__['_fields_']:
-            if k not in ['contact', 'buffer']:
-                state_dict[k] = getattr(self.sim.data, k)
-        return state_dict
-
-    def get_state(self, all_data=False):
-        """Serializes the qpos and qvel state of the MuJoCo emulator."""
-        state = MujocoState.from_mjdata(self.sim.data).flatten()
-        if all_data:
-            full_state = self.get_full_state()
-            radius = self.get_radius()
-            return (state, full_state, radius)
-        return state
-
-    def get_radius(self):
-        return self.env.env.RADIUS
-
-    def set_state(self, x, sim_data=None, radius=None, forward=True):
-        """Restores qpos and qvel, calling forward() to derive other values."""
-        state = MujocoState.from_flattened(x, self.sim)
-        if radius is not None:
-            self.set_radius(radius)
-        state.set_mjdata(self.sim.data, old_mujoco=True)
-        if sim_data is not None:
-            # set more than just qacc, qvel, qpos
-            self.set_arbitrary_state(sim_data)
-        if forward:
-            self.sim.model.forward()  # put mjData in consistent state
-
-    def set_arbitrary_state(self, sim_data):
-        for k, v in sim_data.items():
-            setattr(self.sim.data, k, v)
-
-    def set_radius(self, r):
-        self.env.env.RADIUS = r
-        self.env.env._set_geom_radius()
-
-    def reset(self):
-        """See base class."""
-        return self.env.reset()
-
-    def step(self, a):
-        """See base class."""
-        return self.env.step(a)
+from aprl.common.mujoco import MujocoState, ResettableEnv
 
 
 class MujocoResettableWrapper(ResettableEnv, gym.Wrapper):
